Fix create_schema.sql script

+++ src/main/resources/create_schema.sql
@@ -3,7 +3,7 @@ CREATE DATABASE IF NOT EXISTS phonebook_database;
 CREATE SEQUENCE personal_data_sequence;
 
 CREATE TABLE personal_data (
-    id INTEGER PRIMARY KEY DEFAULT NEXTVAL('personal_data_sequence'),
+    id BIGINT PRIMARY KEY DEFAULT NEXTVAL('personal_data_sequence'),
     name CHARACTER VARYING (64),
     surname CHARACTER VARYING (64)
 );
@@ -11,14 +11,14 @@ CREATE TABLE personal_data (
 CREATE SEQUENCE phonebook_entry_sequence;
 
 CREATE TABLE phonebook_entry (
-    id INTEGER PRIMARY KEY DEFAULT NEXTVAL('phonebook_entry_sequence'),
+    id BIGINT PRIMARY KEY DEFAULT NEXTVAL('phonebook_entry_sequence'),
     personal_data_id INTEGER UNIQUE NOT NULL REFERENCES personal_data(id)
 );
 
 CREATE SEQUENCE phone_number_sequence;
 
 CREATE TABLE phone_number (
-    id INTEGER PRIMARY KEY DEFAULT NEXTVAL('phone_number_sequence'),
+    id BIGINT PRIMARY KEY DEFAULT NEXTVAL('phone_number_sequence'),
     number CHARACTER VARYING (64) UNIQUE,
     phone_number_type CHARACTER (4),
     phonebook_entry_id INTEGER NOT NULL REFERENCES phonebook_entry(id)
